fix: optimize background creation and design
Prior to this fix, if the user wants to change the box size to a smaller number, they will have performance issues. For example, if the box size has been set to one, each frame will take between 800 and 1000 ms to draw
Now this operation is done in less than 1ms!